Refactored frontend
