+++ frontend/src/components/SessionErrorComponent.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { ErrorComponent, ErrorRouteProps } from "@tanstack/react-router";
-import { SessionNotFoundError } from "../sessions";
+import { SessionNotFoundError } from "../sessionsService";
 
 export function SessionErrorComponent({ error }: ErrorRouteProps) {
   if (error instanceof SessionNotFoundError) {
